WIP: Add sensor pipeline
